Feat(repository): update ProductRepository

+++ src/main/java/com/project/api_crud/repository/ProductRepository.java
@@ -2,11 +2,13 @@ package com.project.api_crud.repository;
 
 import com.project.api_crud.model.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.math.BigDecimal;
 import java.util.Optional;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
 
     Optional<Product> findByName(String name);
 
-    Optional<Product> findByPrice(Integer price);
+    Optional<Product> findByPrice(BigDecimal price);
 }
